+++ app/hertz_gateway/utils/function.go
@@ -2,9 +2,9 @@ package utils
 
 import "context"
 
-func GetUserIdFromCtx(ctx context.Context) int32 {
+func GetUserIdFromCtx(ctx context.Context) int64 {
 	if ctx.Value(UserIdKey) == nil {
 		return 0
 	}
-	return ctx.Value(UserIdKey).(int32)
+	return ctx.Value(UserIdKey).(int64)
 }
